requirements.txt
uvicorn
sqlmodel
pymsql
bcrypt
python-jose
python-dotenv
